Implement numeric instructions and fix examples

+++ examples/addative_hello_world.rs
@@ -23,7 +23,7 @@ fn main() {
             ResultType::empty(),
             vec![
                 0i32.into(),
-                message.len().into(),
+                (message.len() as u32).into(),
                 ControlInstruction::Call(print_function).into(),
             ]
             .into(),
